moves unit tests to their respective packages
